coding refacto

+++ disc.go
@@ -1,6 +1,6 @@
 package main
 
-func square(b, h int) (a int) {
+func square(b, h float64) (a float64) {
 	a = b
 	for ; h > 1; h-- {
 		a = a * b
@@ -8,6 +8,6 @@ func square(b, h int) (a int) {
 	return
 }
 
-func disc(b, a, c int) (dix int) {
-	return square(b, 2) - (4 * a * c)
+func disc(b, a, c float64) (dix float64) {
+	return square(b, 2.0) - (4 * a * c)
 }
